refactor: move issue parsing logic to bot module
