+++ modules/bot/package.json
@@ -12,14 +12,21 @@
   },
   "dependencies": {
     "@electron/bugbot-shared": "*",
+    "mdast-util-from-markdown": "^0.8.5",
+    "mdast-util-heading-range": "^2.1.5",
+    "mdast-util-to-string": "^2.0.0",
     "node-fetch": "^2.6.1",
     "probot": "^11.3.0",
+    "semver": "^7.3.5",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
     "@types/jest": "^26.0.23",
+    "@types/mdast": "^3.0.3",
     "@types/node": "^15.3.0",
     "@types/node-fetch": "^2.5.10",
+    "@types/semver": "^7.3.6",
+    "@types/unist": "^2.0.3",
     "jest": "^26.6.3",
     "nock": "^13.0.11",
     "smee-client": "^1.2.2"
