+++ modules/shared/package.json
@@ -8,17 +8,9 @@
     "test": "jest --runInBand",
     "test:ci": "jest --runInBand --coverage"
   },
-  "dependencies": {
-    "mdast-util-from-markdown": "^0.8.5",
-    "mdast-util-heading-range": "^2.1.5",
-    "mdast-util-to-string": "^2.0.0",
-    "semver": "^7.3.5"
-  },
+  "dependencies": {},
   "devDependencies": {
     "@types/jest": "^26.0.23",
-    "@types/mdast": "^3.0.3",
-    "@types/semver": "^7.3.6",
-    "@types/unist": "^2.0.3",
     "jest": "^26.6.3",
     "ts-jest": "^26.5.6"
   }
